fix: moved all files in Editor-assembly
