+++ Assets/Scripts/Editor/AnimationCurveBase.cs
@@ -3,7 +3,7 @@ using System.Collections.Generic;
 using UnityEditor;
 using UnityEngine;
 
-namespace Citrine.Utils.AnimationCompression
+namespace Citrine.Utils.Editor.AnimationCompression
 {
     internal abstract class AnimationCurveBase<T> where T : struct
     {
