+++ Assets/Scripts/Editor/AnimationCurveBase.cs.meta
