+++ Assets/Scripts/Editor/KeyframeReducer.cs.meta
