+++ Assets/Scripts/Editor/QuaternionAnimationCurve.cs.meta
