+++ Assets/Scripts/Editor/Vector3AnimationCurve.cs.meta
